refactor(#8/post): 리포지토리 인터페이스를 infrastructure/repository로 위치 변경
