+++ src/main/java/org/sopt/domain/post/infrastructure/repository/PostRepository.java
@@ -1,4 +1,4 @@
-package org.sopt.domain.post.infrastructure;
+package org.sopt.domain.post.infrastructure.repository;
 
 import java.util.List;
 import java.util.Optional;
